backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
requests==2.32.3
python-dateutil==2.9.0.post0
django-axes==6.5.0
selenium==4.22.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
bcrypt==4.1.3